FlaskWebAPI/requirements.txt
flask
flask_cors
pyjwt
cryptography
requests
azure-storage-blob